Move tests depending on manual preparation to `xt/`
The test `02-installation.t` depends on having the module `SVG::Plot`
installed. That module was listed in `META6.json` `test-depends`, but this
did not help, as the module won't be installed until after installation.
So to sucessfully run the test one needs to manually install `SVG::Plot`
beforehand. So this test is no good fit for being run during automated
testing during installation.