Refactor logic to use CUR apis
Code interacting with modules should do so through the CompUnit::Repository interface, and not rely on implementation details such as the directory layout of a repository. This refactors the matching and loading logic to be CUR agnostic other than `$repo ~~ CompUnit::Repository::FileSystem ?? $repo.distribution !! $repo.installed;` (i.e. this is no universal "give me all distributions of whatever CUR this is" interface yet), which should fix issues like testing a consumer of Pluggable working with `-Ilib` but not `-I.`.